+++ backend/apid/graphql/entity.go
@@ -32,6 +32,12 @@ func (*entityImpl) ID(p graphql.ResolveParams) (string, error) {
 	return globalid.EntityTranslator.EncodeToString(p.Source), nil
 }
 
+// Metadata implements response to request for 'metadata' field.
+func (*entityImpl) Metadata(p graphql.ResolveParams) (interface{}, error) {
+	entity := p.Source.(*corev2.Entity)
+	return entity.GetRedactedEntity(), nil
+}
+
 // ExtendedAttributes implements response to request for 'extendedAttributes' field.
 func (*entityImpl) ExtendedAttributes(p graphql.ResolveParams) (interface{}, error) {
 	entity := p.Source.(*corev2.Entity)