+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 highlighted option, instead of nagios-perfdata.
 - Fixed a bug where silences would not expire on event resolution.
 
+- [GraphQL] Ensures that proxy entity label & annotations are redacted.
+
 ## [5.12.0] - 2019-08-22
 
 ### Added